fix some issues
