+++ docs/en/setup/service-agent/virtual-genai.md
@@ -1,6 +1,6 @@
 # Virtual GenAI
 
-Virtual cache represent the Generative AI service nodes detected by [server agents' plugins](server-agents.md). The performance
+Virtual GenAI represent the Generative AI service nodes detected by [server agents' plugins](server-agents.md). The performance
 metrics of the GenAI operations are also from the GenAI client-side perspective.
 
 For example, an Spring-ai plugin in the Java agent could detect the latency of a chat completion request.
@@ -13,4 +13,5 @@ The GenAI operation span should have
 - Tag key = `gen_ai.response.model`, value = The name of the GenAI model a response is being made to, e.g. gpt-4o, claude-3-5-sonnet
 - Tag key = `gen_ai.usage.input_tokens`, value = The number of tokens used in the GenAI input (prompt)
 - Tag key = `gen_ai.usage.output_tokens`, value = The number of tokens used in the GenAI response (completion)
+- Tag key = `gen_ai.server.time_to_first_token`, value = The duration in milliseconds from the start of the request until the first token is received. Note: This metric is only available for streaming requests.
 - If the GenAI service is a remote API (e.g. OpenAI), the span's peer would be the network address (IP or domain) of the GenAI server.
