+++ docs/menu.yml
@@ -156,6 +156,10 @@ catalog:
                 path: "/en/setup/backend/dashboards-so11y-java-agent"
               - name: "SkyWalking Go Agent self telemetry"
                 path: "/en/setup/backend/dashboards-so11y-go-agent"
+          - name: "GenAI"
+            catalog:
+              - name: "Virtual Genai"
+                path: "/en/setup/service-agent/virtual-genai"
       - name: "Configuration Vocabulary"
         path: "/en/setup/backend/configuration-vocabulary"
       - name: "Advanced Setup"
