+++ oap-server/analyzer/gen-ai-analyzer/pom.xml
@@ -25,7 +25,7 @@
     </parent>
     <modelVersion>4.0.0</modelVersion>
 
-    <artifactId>genAI-analyzer</artifactId>
+    <artifactId>gen-ai-analyzer</artifactId>
 
     <dependencies>
         <dependency>
