+++ oap-server/analyzer/gen-ai-analyzer/src/main/resources/META-INF/services/org.apache.skywalking.oap.server.library.module.ModuleDefine
@@ -16,4 +16,4 @@
 #
 #
 
-org.apache.skywalking.oap.meter.analyzer.module.GenAIAnalyzerModule
+org.apache.skywalking.oap.analyzer.genai.module.GenAIAnalyzerModule
