+++ oap-server/analyzer/gen-ai-analyzer/src/main/resources/META-INF/services/org.apache.skywalking.oap.server.library.module.ModuleProvider
@@ -15,4 +15,4 @@
 # limitations under the License.
 #
 
-org.apache.skywalking.oap.meter.analyzer.GenAIAnalyzerModuleProvider
+org.apache.skywalking.oap.analyzer.genai.GenAIAnalyzerModuleProvider
