+++ oap-server/server-starter/src/main/resources/oal/virtual-gen-ai.oal
@@ -20,13 +20,16 @@ gen_ai_provider_resp_time = from(GenAIProviderAccess.latency).longAvg();
 gen_ai_provider_sla = from(GenAIProviderAccess.*).percent(status == true);
 gen_ai_provider_cpm = from(GenAIProviderAccess.*).cpm();
 
-gen_ai_provider_latency_avg = from(GenAIProviderAccess.latency).longAvg();
-gen_ai_provider_latency_percentile = from(GenAIProviderAccess.latency).percentile(10);
+gen_ai_provider_latency_percentile = from(GenAIProviderAccess.latency).percentile2(10);
 
 gen_ai_provider_input_tokens_sum = from(GenAIProviderAccess.inputTokens).sum();
+gen_ai_provider_input_tokens_avg = from(GenAIProviderAccess.inputTokens).longAvg();
+
 gen_ai_provider_output_tokens_sum = from(GenAIProviderAccess.outputTokens).sum();
+gen_ai_provider_output_tokens_avg = from(GenAIProviderAccess.outputTokens).longAvg();
 
 gen_ai_provider_total_cost = from(GenAIProviderAccess.totalCost).sum();
+gen_ai_provider_avg_cost = from(GenAIProviderAccess.totalCost).doubleAvg();
 
 gen_ai_model_call_cpm = from(GenAIModelAccess.*).cpm();
 
@@ -34,12 +37,15 @@ gen_ai_model_sla = from(GenAIModelAccess.*).percent(status == true);
 
 gen_ai_model_latency_avg = from(GenAIModelAccess.latency).longAvg();
 
-gen_ai_model_latency_percentile = from(GenAIModelAccess.latency).percentile(10);
+gen_ai_model_latency_percentile = from(GenAIModelAccess.latency).percentile2(10);
 
 gen_ai_model_ttft_avg = from(GenAIModelAccess.timeToFirstToken).filter(timeToFirstToken > 0).longAvg();
-gen_ai_model_ttft_percentile = from(GenAIModelAccess.timeToFirstToken).filter(timeToFirstToken > 0).percentile(10);
+gen_ai_model_ttft_percentile = from(GenAIModelAccess.timeToFirstToken).filter(timeToFirstToken > 0).percentile2(10);
 
 gen_ai_model_input_tokens_sum = from(GenAIModelAccess.inputTokens).sum();
+gen_ai_model_input_tokens_avg = from(GenAIModelAccess.inputTokens).longAvg();
 gen_ai_model_output_tokens_sum = from(GenAIModelAccess.outputTokens).sum();
+gen_ai_model_output_tokens_avg = from(GenAIModelAccess.outputTokens).longAvg();
 
 gen_ai_model_total_cost = from(GenAIModelAccess.totalCost).sum();
+gen_ai_model_avg_cost = from(GenAIModelAccess.totalCost).doubleAvg();
