+++ oap-server/server-starter/src/main/resources/ui-initialized-templates/menu.yaml
