+++ oap-server/server-starter/src/main/resources/ui-initialized-templates/virtual_genai/virtual-genai-model.json
@@ -107,9 +107,9 @@
                 },
                 {
                   "x": 12,
-                  "y": 0,
+                  "y": 24,
                   "w": 12,
-                  "h": 11,
+                  "h": 12,
                   "i": "1",
                   "type": "Widget",
                   "expressions": [
@@ -136,9 +136,9 @@
                 },
                 {
                   "x": 0,
-                  "y": 11,
+                  "y": 24,
                   "w": 12,
-                  "h": 11,
+                  "h": 12,
                   "i": "0",
                   "type": "Widget",
                   "expressions": [
@@ -165,7 +165,7 @@
                 },
                 {
                   "x": 12,
-                  "y": 11,
+                  "y": 36,
                   "w": 12,
                   "h": 11,
                   "i": "5",
@@ -192,37 +192,8 @@
                   "moved": false
                 },
                 {
-                  "x": 0,
-                  "y": 22,
-                  "w": 12,
-                  "h": 12,
-                  "i": "14",
-                  "type": "Widget",
-                  "expressions": [
-                    "gen_ai_model_latency_avg"
-                  ],
-                  "graph": {
-                    "type": "Line",
-                    "showXAxis": true,
-                    "showYAxis": true
-                  },
-                  "widget": {
-                    "name": "AvgLatency",
-                    "title": "Average Latency",
-                    "tips": "The average latency of model access."
-                  },
-                  "metricConfig": [
-                    {
-                      "label": "Avg Latency",
-                      "unit": "ms"
-                    }
-                  ],
-                  "id": "0-0-14",
-                  "moved": false
-                },
-                {
                   "x": 12,
-                  "y": 22,
+                  "y": 47,
                   "w": 12,
                   "h": 12,
                   "i": "13",
@@ -244,9 +215,9 @@
                 },
                 {
                   "x": 0,
-                  "y": 34,
+                  "y": 36,
                   "w": 12,
-                  "h": 10,
+                  "h": 11,
                   "i": "20",
                   "type": "Widget",
                   "widget": {
@@ -278,10 +249,10 @@
                   ]
                 },
                 {
-                  "x": 12,
-                  "y": 34,
+                  "x": 0,
+                  "y": 47,
                   "w": 12,
-                  "h": 10,
+                  "h": 12,
                   "i": "21",
                   "type": "Widget",
                   "widget": {
@@ -303,6 +274,108 @@
                   "moved": false,
                   "expressions": [
                     "gen_ai_model_ttft_percentile"
+                  ],
+                  "typesOfMQE": [
+                    "TIME_SERIES_VALUES"
+                  ]
+                },
+                {
+                  "x": 0,
+                  "y": 11,
+                  "w": 12,
+                  "h": 13,
+                  "i": "22",
+                  "type": "Widget",
+                  "widget": {
+                    "name": "AvgInputTokens",
+                    "title": "Average Input Tokens",
+                    "tips": "The average number of input tokens used per model call."
+                  },
+                  "metricConfig": [
+                    {
+                      "label": "Avg Input Tokens"
+                    }
+                  ],
+                  "graph": {
+                    "type": "Line",
+                    "step": false,
+                    "smooth": false,
+                    "showSymbol": true,
+                    "showXAxis": true,
+                    "showYAxis": true
+                  },
+                  "id": "0-0-22",
+                  "moved": false,
+                  "expressions": [
+                    "gen_ai_model_input_tokens_avg"
+                  ],
+                  "typesOfMQE": [
+                    "TIME_SERIES_VALUES"
+                  ]
+                },
+                {
+                  "x": 12,
+                  "y": 11,
+                  "w": 12,
+                  "h": 13,
+                  "i": "23",
+                  "type": "Widget",
+                  "widget": {
+                    "name": "AvgOutputTokens",
+                    "title": "Average Output Tokens",
+                    "tips": "The average number of output tokens used per model call."
+                  },
+                  "metricConfig": [
+                    {
+                      "label": "Avg Output Tokens"
+                    }
+                  ],
+                  "graph": {
+                    "type": "Line",
+                    "step": false,
+                    "smooth": false,
+                    "showSymbol": true,
+                    "showXAxis": true,
+                    "showYAxis": true
+                  },
+                  "id": "0-0-23",
+                  "moved": false,
+                  "expressions": [
+                    "gen_ai_model_output_tokens_avg"
+                  ],
+                  "typesOfMQE": [
+                    "TIME_SERIES_VALUES"
+                  ]
+                },
+                {
+                  "x": 12,
+                  "y": 0,
+                  "w": 12,
+                  "h": 11,
+                  "i": "24",
+                  "type": "Widget",
+                  "widget": {
+                    "name": "AvgCost",
+                    "title": "Average Cost",
+                    "tips": "The average cost of model calls."
+                  },
+                  "metricConfig": [
+                    {
+                      "label": "Avg Cost"
+                    }
+                  ],
+                  "graph": {
+                    "type": "Line",
+                    "step": false,
+                    "smooth": false,
+                    "showSymbol": true,
+                    "showXAxis": true,
+                    "showYAxis": true
+                  },
+                  "id": "0-0-24",
+                  "moved": false,
+                  "expressions": [
+                    "gen_ai_model_avg_cost"
                   ],
                   "typesOfMQE": [
                     "TIME_SERIES_VALUES"
