+++ oap-server/server-starter/src/main/resources/ui-initialized-templates/virtual_genai/virtual-genai-root.json
@@ -55,3 +55,4 @@
     }
   }
 ]
+
