+++ test/e2e-v2/cases/storage/expected/config-dump.yml
@@ -38,6 +38,7 @@ core.default.enableDataKeeperExecutor=true
 agent-analyzer.default.slowCacheReadThreshold=default:20,redis:10
 receiver-ebpf.default.continuousPolicyCacheTimeout=60
 receiver-ebpf.default.gRPCSslKeyPath=
+gen-ai-analyzer.provider=default
 receiver-browser.provider=default
 agent-analyzer.default.segmentStatusAnalysisStrategy=FROM_SPAN_STATUS
 envoy-metric.default.maxConcurrentCallsPerConnection=0
@@ -64,7 +65,6 @@ telemetry.provider=prometheus
 core.default.trainingPeriodHttpUriRecognitionPattern=60
 promql.default.restContextPath=/
 core.default.maxHeapMemoryUsagePercent=96
-genAI-analyzer.provider=default
 aws-firehose.default.contextPath=/
 agent-analyzer.default.slowCacheWriteThreshold=default:20,redis:10
 envoy-metric.default.maxMessageSize=0
