+++ test/e2e-v2/cases/virtual-genai/docker-compose.yml
