+++ test/e2e-v2/cases/virtual-genai/e2e.yaml
@@ -42,3 +42,4 @@ verify:
   cases:
     - includes:
         - ./virtual-genai.yaml
+
