+++ test/e2e-v2/cases/virtual-genai/expected/instance.yml
@@ -15,7 +15,7 @@
 
 {{- contains . }}
 - id: {{ notEmpty .id }}
-  name: gpt-4.1-mini
+  name: gpt-4.1-mini-2025-04-14
   instanceuuid: {{ notEmpty .instanceuuid }}
   attributes: []
   language: UNKNOWN
