+++ test/e2e-v2/cases/virtual-genai/expected/metrics-has-value.yml
