+++ test/e2e-v2/cases/virtual-genai/expected/service.yml
