+++ test/e2e-v2/java-test-service/e2e-service-provider/pom.xml
@@ -112,11 +112,6 @@
             <version>23.0</version>
         </dependency>
 
-        <dependency>
-            <groupId>com.alibaba</groupId>
-            <artifactId>fastjson</artifactId>
-            <version>1.2.83</version>
-        </dependency>
     </dependencies>
 
     <build>
